requirements/development.txt
-r base.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1            # Test fixtures / model factories

# Code quality
ruff==0.9.1                   # Fast linter + formatter
mypy==1.14.1
django-stubs==5.1.1

# Development utilities
ipython==8.31.0
Werkzeug==3.1.3               # For runserver_plus with SSL
